AJAX requests, Merged templates and Tracking function modified
Changed the behaviour so that we don’t reload the page when we click
« point! » by using AJAX.

Also, I merged index.html and layout.html and got rid of index.html.

Finally, I modified the way I call the tracking function. It now calls
the astropointer.py with extra parameters in a different forked
process. This works once, but won’t work when we change object. THIS
NEEDS TO BE MODIFIED.